# integrations/anim-works/openapi.yaml
openapi: 3.1.0
info:
  title: Animation Reference Partner API
  version: "1.0.0"
  description: |
    Read-only access to the Animation Reference library, plus the handoff
    exchange used when a user sends a clip from animationreference.org into a
    partner tool.

    All endpoints require an API key in the `X-API-Key` header. Keys are
    server-side credentials: do not ship them to a browser.
  contact:
    email: michaelfred124@gmail.com
  termsOfService: https://animationreference.org/terms
servers:
  - url: https://animationreference.org/api/v1
security:
  - ApiKeyAuth: []
paths:
...
  /tags:
    get:
      operationId: listTags
      summary: Tags ordered by clip count
      parameters:
        - name: q
          in: query
          description: Substring filter on the tag name
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: One page of tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TagSummary" }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    TagSummary:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        videoCount: { type: integer }
        url: { type: string, format: uri }
...
    Pagination:
      type: object
      properties:
        limit: { type: integer }
        total: { type: integer }
        nextCursor:
          type: [string, "null"]
          description: Opaque. Pass back unchanged; never parse or construct one.
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryAfter: { type: integer, description: Seconds; on 429 only }
  responses:
...
    Unauthorized:
      description: Missing, invalid, revoked or expired key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...